feat: get response from API and display on UI as bot's message (#25)
